refactor(common): FE-432 Remove components/form barrel file
