+++ apps/storefront/src/components/B3CustomForm.tsx
@@ -1,19 +1,17 @@
 import { Grid } from '@mui/material';
 
+import { B2BControlMultiTextField } from './form/B2BControlMultiTextField';
+import { B3ControlAutocomplete } from './form/B3ControlAutocomplete';
+import { B3ControlCheckbox } from './form/B3ControlCheckbox';
+import { B3ControlFileUpload } from './form/B3ControlFileUpload';
+import { B3ControlPicker } from './form/B3ControlPicker';
+import { B3ControlProductRadio } from './form/B3ControlProductRadio';
+import { B3ControlRadioGroup } from './form/B3ControlRadioGroup';
+import { B3ControlRectangle } from './form/B3ControlRectangle';
+import { B3ControlSelect } from './form/B3ControlSelect';
+import { B3ControlSwatchRadio } from './form/B3ControlSwatchRadio';
+import { B3ControlTextField } from './form/B3ControlTextField';
 import B3UI from './form/ui';
-import {
-  B2BControlMultiTextField,
-  B3ControlAutocomplete,
-  B3ControlCheckbox,
-  B3ControlFileUpload,
-  B3ControlPicker,
-  B3ControlProductRadio,
-  B3ControlRadioGroup,
-  B3ControlRectangle,
-  B3ControlSelect,
-  B3ControlSwatchRadio,
-  B3ControlTextField,
-} from './form';
 
 export default function B3CustomForm(props: B3UI.B3CustomFormProps) {
   const { formFields, errors, control, getValues, setValue, setError } = props;
