+++ apps/storefront/src/components/form/B3ControlProductRadio.tsx
@@ -1,6 +1,6 @@
 import { PRODUCT_DEFAULT_IMAGE } from '@/constants';
 
-import B3ControlRadioGroup from './B3ControlRadioGroup';
+import { B3ControlRadioGroup } from './B3ControlRadioGroup';
 import { ProductImageContainer } from './styled';
 import Form from './ui';
 
@@ -9,7 +9,7 @@ const NoneOption = {
   value: '',
 };
 
-export default function B3ControlProductRadio(props: Form.B3UIProps) {
+export function B3ControlProductRadio(props: Form.B3UIProps) {
   const { options } = props;
 
   const getProductImageUrl = (url = '') => url.replace('{:size}', 'original');
