+++ apps/storefront/src/components/form/B3ControlRectangle.tsx
@@ -6,7 +6,7 @@ import { useB3Lang } from '@/lib/lang';
 import { StyleRectangleFormControlLabel } from './styled';
 import Form from './ui';
 
-export default function B3ControlRectangle({ control, errors, ...rest }: Form.B3UIProps) {
+export function B3ControlRectangle({ control, errors, ...rest }: Form.B3UIProps) {
   const {
     fieldType,
     name,
