--- apps/storefront/src/components/form/index.ts
@@ -1,12 +1,0 @@
-export { Captcha } from '../captcha/Captcha';
-export { default as B3ControlCheckbox } from './B3ControlCheckbox';
-export { default as B3ControlFileUpload } from './B3ControlFileUpload';
-export { default as B3ControlPicker } from './B3ControlPicker';
-export { default as B3ControlProductRadio } from './B3ControlProductRadio';
-export { default as B3ControlRadioGroup } from './B3ControlRadioGroup';
-export { default as B3ControlRectangle } from './B3ControlRectangle';
-export { default as B3ControlSelect } from './B3ControlSelect';
-export { default as B3ControlSwatchRadio } from './B3ControlSwatchRadio';
-export { default as B3ControlTextField } from './B3ControlTextField';
-export { default as B3ControlAutocomplete } from './B3ControlAutocomplete';
-export { default as B2BControlMultiTextField } from './B2BControlMultiTextField';
