+++ apps/storefront/src/pages/ForgotPassword/index.tsx
@@ -5,7 +5,7 @@ import { Box, ImageListItem, Typography } from '@mui/material';
 
 import { B3Card, B3CustomForm } from '@/components';
 import CustomButton from '@/components/button/CustomButton';
-import { Captcha } from '@/components/form';
+import { Captcha } from '@/components/captcha/Captcha';
 import B3Spin from '@/components/spin/B3Spin';
 import { useMobile } from '@/hooks';
 import { useB3Lang } from '@/lib/lang';
