+++ apps/storefront/src/pages/QuickOrder/components/QuickAdd.tsx
@@ -3,7 +3,7 @@ import { useForm } from 'react-hook-form';
 import { Box, Grid, Typography } from '@mui/material';
 
 import CustomButton from '@/components/button/CustomButton';
-import B3ControlTextField from '@/components/form/B3ControlTextField';
+import { B3ControlTextField } from '@/components/form/B3ControlTextField';
 import B3Spin from '@/components/spin/B3Spin';
 import { CART_URL } from '@/constants';
 import { useBlockPendingAccountViewPrice, useFeatureFlags } from '@/hooks';
